+++ astrbot/core/pipeline/process_stage/utils.py
@@ -15,13 +15,15 @@ async def inject_kb_context(
         p_ctx: Pipeline context
         req: Provider request
     """
-    kb_injector = p_ctx.plugin_manager.context.kb_manager.get_kb_injector()
-    if not kb_injector:
+    kb_mgr = p_ctx.plugin_manager.context.kb_manager
+    kb_names = p_ctx.astrbot_config.get("kb_names", [])
+
+    if not kb_names:
         return
-    kb_context = await kb_injector.retrieve_and_inject(
-        unified_msg_origin=umo,
+
+    kb_context = await kb_mgr.retrieve(
         query=req.prompt,
-        top_k=top_k,
+        kb_names=kb_names,
     )
     if not kb_context:
         return
